refactor parser de conversao
