+++ api/core/dao/geosampa/geosampa_address_query.py
@@ -1,7 +1,7 @@
 from core.integrations import geosampa
 from core.utils.geo import convert_points_to_sirgas
 from ..parsers.geosampa_geojson import GeosampaGeoJsonParser
-from ..convert_to_wgs_84 import WGS84_conversor
+from ..parsers.convert_to_wgs_84 import WGS84_conversor
 from ..camadas_search import DetailCamada
 
 from core.schemas.camadas import CamadaParamInternal
